requirements.txt
streamlit==1.36.0
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
matplotlib==3.8.4
